+++ src/main/java/com/mobble/mobbleserver/application/notification/device/service/DeviceTokenService.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.notification.device.service;
+package com.mobble.mobbleserver.application.notification.device.service;
 
 import com.mobble.mobbleserver.application.member.port.required.MemberReadPort;
 import com.mobble.mobbleserver.domain.member.Member;
@@ -6,20 +6,20 @@ import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.member.MemberErrorCode;
 import com.mobble.mobbleserver.domain.notification.device.DeviceToken;
 import com.mobble.mobbleserver.domain.notification.device.Platform;
-import com.mobble.mobbleserver.infrastructure.persistence.notification.device.DeviceTokenRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.notification.device.JpaDeviceTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Objects;
 
-import static com.mobble.mobbleserver.refactor.notification.core.dto.NotificationDto.RegisterTokenReq;
+import static com.mobble.mobbleserver.infrastructure.web.notification.dto.NotificationDto.RegisterTokenReq;
 
 @Service
 @RequiredArgsConstructor
 public class DeviceTokenService {
 
-    private final DeviceTokenRepository deviceTokenRepository;
+    private final JpaDeviceTokenRepository jpaDeviceTokenRepository;
 
     private final MemberReadPort memberReadPort;
 
@@ -28,19 +28,19 @@ public class DeviceTokenService {
         Member member = findMemberByMemberIdOrThrow(memberId);
         Platform platform = Platform.valueOf(request.platform().toUpperCase());
 
-        return deviceTokenRepository.findByToken(request.token())
+        return jpaDeviceTokenRepository.findByToken(request.token())
                 .map(deviceToken -> {
                     deviceToken.reassignTo(member);
                     return deviceToken;
                 })
                 .orElseGet(() ->
-                        deviceTokenRepository.save(DeviceToken.create(member, request.token(), platform))
+                        jpaDeviceTokenRepository.save(DeviceToken.create(member, request.token(), platform))
                 );
     }
 
     @Transactional
     public void disable(Long memberId, Long deviceTokenId) {
-        DeviceToken deviceToken = deviceTokenRepository.findById(deviceTokenId).orElseThrow();
+        DeviceToken deviceToken = jpaDeviceTokenRepository.findById(deviceTokenId).orElseThrow();
 
         if (!Objects.equals(deviceToken.getMember().getId(), memberId)) throw new IllegalStateException("forbidden");
